+++ src/test/data/object/formatTree.js
@@ -11,7 +11,14 @@ import formatTree from './../../../data/object/formatTree'
     {
       pid: 0,
       id: 2,
-      name: '2'
+      name: '2',
+      children: [
+        {
+          pid: 2,
+          id: 100,
+          name: '2-100'
+        }
+      ]
     },
     {
       pid: 0,
@@ -40,9 +47,23 @@ import formatTree from './../../../data/object/formatTree'
   if (treeList.length != 2 || treeList[0].name != '2' || treeList[1].name != '1') {
     console.error('formatTree未成功,长度或者顺序错误')
   } else {
-    let insideData = treeList[0].children[0].children[0]
+    let insideData = treeList[0].children[1].children[0]
     if (!insideData || insideData.name != '3-6') {
       console.error('formatTree未成功，内部值错误')
     }
+  }
+
+  let treeList2 = formatTree(list, {
+    parentId: 'pid',
+    childrenFormat: function(childrenList, originItem) {
+      if (childrenList) {
+        childrenList = []
+      }
+      return childrenList
+    }
+  })
+  let insideData = treeList2[0].children[0].children[0]
+  if (!insideData || insideData.name != '3-6') {
+    console.error('formatTree未成功，childrenFormat错误')
   }
 })();